ORCID honours the state param, so no need to ignore it
The current ORCID docs [1] say:

> Appended to the end of that link will be a six-digit authorization
> code and any state parameter you specified

We've tested this in our system (a rails app at theconversation.com) and
confirmed that sign in via ORCID continues to work when
provider_ignores_state is set to `false`.

oauth2 sign ins are reportedly safer when the state param is checked, so
it may be worth changing the default behaviour.

[1] https://members.orcid.org/api/integrate/orcid-sign-in